Engine: float-facing API (Rand.Range float overload, Frame.Time float)
Students move on to Unity, so the course now writes float everywhere a
decimal number appears. For that to work without teaching casts, the two
API members that hand a number back to student code must return float.

Rand.Range(double, double) becomes Rand.Range(float, float). The double
overload is dropped rather than kept alongside: two overloads would make
Rand.Range(1.5, 2.5) ambiguous-looking to a beginner and would quietly
hand back a double that a float variable cannot hold. A bare-decimal call
is now a compile error, which is the intended outcome.

Frame.Time becomes float (Count / (float)Rate). The cast is engine code.

Screen's parameters and the DrawCommand/TextCanvas/GameSession internals
stay double on purpose: a float argument converts implicitly, so student
code never casts, and the renderer's maths is unchanged - which is why
every frame golden still reproduces bit-for-bit.

Test snippets written in student style switch to float too, so the tests
reflect what students actually write. SourceCompilerTests' representative
beginner program now uses MathF.Sqrt, which also proves MathF is present
in the shipped facades.

+++ engine/Harness/Game/Frame.cs
@@ -5,5 +5,5 @@ public static class Frame
 {
     public const int Rate = 30;
     public static int Count => GameHost.State.FrameCount;
-    public static double Time => Count / (double)Rate;
+    public static float Time => Count / (float)Rate;
 }
